Compiling and rearranged but tests not working
