+++ fitsio/src/types.rs
@@ -1,5 +1,4 @@
 //! Data types used within `fitsio`
-use descriptions::*;
 
 /// Enumeration of different data types used for column and key types
 #[allow(missing_docs)]
@@ -23,173 +22,6 @@ pub enum DataType {
     TCOMPLEX,
     TDBLCOMPLEX,
 }
-
-macro_rules! datatype_into_impl {
-    ($t: ty) => (
-        impl From<DataType> for $t {
-            fn from(original: DataType) -> $t {
-                match original {
-                    DataType::TBIT => 1,
-                    DataType::TBYTE => 11,
-                    DataType::TSBYTE => 12,
-                    DataType::TLOGICAL => 14,
-                    DataType::TSTRING => 16,
-                    DataType::TUSHORT => 20,
-                    DataType::TSHORT => 21,
-                    DataType::TUINT => 30,
-                    DataType::TINT => 31,
-                    DataType::TULONG => 40,
-                    DataType::TLONG => 41,
-                    DataType::TLONGLONG => 81,
-                    DataType::TFLOAT => 42,
-                    DataType::TDOUBLE => 82,
-                    DataType::TCOMPLEX => 83,
-                    DataType::TDBLCOMPLEX => 163,
-                }
-            }
-        }
-    )
-}
-
-datatype_into_impl!(u8);
-datatype_into_impl!(i32);
-datatype_into_impl!(u32);
-datatype_into_impl!(i64);
-datatype_into_impl!(u64);
-
-/// Data types used for defining images
-#[allow(missing_docs)]
-#[derive(Debug, PartialEq, Eq, Clone, Copy)]
-pub enum ImageType {
-    UnsignedByte,
-    Byte,
-    Short,
-    UnsignedShort,
-    Long,
-    UnsignedLong,
-    LongLong,
-    Float,
-    Double,
-}
-
-macro_rules! imagetype_into_impl {
-    ($t: ty) => (
-        impl From<ImageType> for $t {
-            fn from(original: ImageType) -> $t {
-                match original {
-                    ImageType::UnsignedByte => 8,
-                    ImageType::Byte => 10,
-                    ImageType::Short => 16,
-                    ImageType::UnsignedShort => 20,
-                    ImageType::Long => 32,
-                    ImageType::UnsignedLong => 40,
-                    ImageType::LongLong => 64,
-                    ImageType::Float => -32,
-                    ImageType::Double => -64,
-                }
-            }
-        }
-    )
-}
-
-imagetype_into_impl!(i8);
-imagetype_into_impl!(i32);
-imagetype_into_impl!(i64);
-
-/// Description of the current HDU
-///
-/// If the current HDU is an image, then
-/// [`fetch_hdu_info`](struct.FitsFile.html#method.fetch_hdu_info) returns `HduInfo::ImageInfo`.
-/// Otherwise the variant is `HduInfo::TableInfo`.
-#[allow(missing_docs)]
-#[derive(Debug, PartialEq)]
-pub enum HduInfo {
-    ImageInfo {
-        shape: Vec<usize>,
-        image_type: ImageType,
-    },
-    TableInfo {
-        column_descriptions: Vec<ConcreteColumnDescription>,
-        num_rows: usize,
-    },
-    AnyInfo,
-}
-
-macro_rules! hduinfo_into_impl {
-    ($t: ty) => (
-        impl From<HduInfo> for $t {
-            fn from(original: HduInfo) -> $t {
-                match original {
-                    HduInfo::ImageInfo { .. } => 0,
-                    HduInfo::TableInfo { .. } => 2,
-                    HduInfo::AnyInfo => -1,
-                }
-            }
-        }
-    )
-}
-
-hduinfo_into_impl!(i8);
-hduinfo_into_impl!(i32);
-hduinfo_into_impl!(i64);
-
-/// Enumeration of file open modes
-#[allow(missing_docs)]
-#[repr(C)]
-#[derive(Debug, PartialEq, Eq)]
-pub enum FileOpenMode {
-    READONLY,
-    READWRITE,
-}
-
-macro_rules! fileopenmode_into_impl {
-    ($t: ty) => (
-        impl From<FileOpenMode> for $t {
-            fn from(original: FileOpenMode) -> $t {
-                match original {
-                    FileOpenMode::READONLY => 0,
-                    FileOpenMode::READWRITE => 1,
-                }
-            }
-        }
-        )
-}
-
-fileopenmode_into_impl!(u8);
-fileopenmode_into_impl!(u32);
-fileopenmode_into_impl!(u64);
-fileopenmode_into_impl!(i8);
-fileopenmode_into_impl!(i32);
-fileopenmode_into_impl!(i64);
-
-/// Enumeration of options for case sensitivity
-#[allow(missing_docs)]
-#[repr(C)]
-#[derive(Debug, PartialEq, Eq)]
-pub enum CaseSensitivity {
-    CASEINSEN,
-    CASESEN,
-}
-
-macro_rules! casesensitivity_into_impl {
-    ($t: ty) => (
-        impl From<CaseSensitivity> for $t {
-            fn from(original: CaseSensitivity) -> $t {
-                match original {
-                    CaseSensitivity::CASEINSEN => 0,
-                    CaseSensitivity::CASESEN => 1,
-                }
-            }
-        }
-        )
-}
-
-casesensitivity_into_impl!(u8);
-casesensitivity_into_impl!(u32);
-casesensitivity_into_impl!(u64);
-casesensitivity_into_impl!(i8);
-casesensitivity_into_impl!(i32);
-casesensitivity_into_impl!(i64);
 
 #[cfg(test)]
 mod test {
